requirements.txt
# NAI-Auto-Generator requirements

# API 호출 관련
requests>=2.28.0

# 암호화 해시를 위한 argon2 (argon2-cffi)
argon2-cffi>=21.3.0

# 이미지 처리 (Pillow)
Pillow>=9.0.0

# GUI를 위한 PyQt5
PyQt5>=5.15.0

# onnxruntime (V4 호환을 위해 1.13.1 고정)
onnxruntime==1.13.1

# NumPy (onnxruntime 1.13.1 호환을 위해 1.x 버전 사용)
numpy>=1.20,<2

# OS Keyring을 통한 인증 정보 안전 저장 (선택적)
keyring>=23.0.0
